Fix label issue and create book creation module
