+++ data/label.json
@@ -82,5 +82,54 @@
         "label_title" : "War"
       }
     ]
+  },
+  {
+    "id" : 32,
+    "title" : "Kaola",
+    "color" : "red",
+    "items" : [
+      {
+        "id" : 32,
+        "publish_date" : "1111-11-11",
+        "publisher" : "Puto",
+        "cover_state" : "bad",
+        "label_title" : "Kaola"
+      }
+    ]
+  },
+  {
+    "id" : 30,
+    "title" : "KAKIO",
+    "color" : "KAKI",
+    "items" : [
+      {
+        "id" : 50,
+        "publish_date" : "1111-11-11",
+        "publisher" : "KARLOS",
+        "cover_state" : "bad",
+        "label_title" : "KAKIO"
+      }
+    ]
+  },
+  {
+    "id" : 47,
+    "title" : "kalo",
+    "color" : "red",
+    "items" : [
+      {
+        "id" : 4,
+        "publish_date" : "2021-12-12",
+        "publisher" : "ben",
+        "cover_state" : "bad",
+        "label_title" : "kalo"
+      },
+      {
+        "id" : 5,
+        "publish_date" : "2012-11-11",
+        "publisher" : "Conor",
+        "cover_state" : "bad",
+        "label_title" : "kalo"
+      }
+    ]
   }
 ]
